// client/src/pages/GameDetails/style.ts
import styled from "styled-components";

export const Textopronto= styled.div`
  padding: 1rem;
  line-height: 19.2px;
  text-align: left;
  color:#454545;
`

export const Retangulo=styled.div`
  width: 592px;
  height: 130px;
  top: 364px;
  left: 344px;
  gap: 0px;
  border-radius: 4px 0px 0px 0px;
  opacity: 0px;
  background: #FFFFFF;
`